enable hold to select, remove once x1 chip, unify reschedule modal
