+++ src/views/Chores/ArchivedTasks.jsx
@@ -432,6 +432,16 @@ const ArchivedTasks = () => {
     setSelectedChores(newSelection)
   }
 
+  // Press-and-hold on a task card enters multi-select with that task picked
+  const enterMultiSelectWithChore = choreId => {
+    if (!isMultiSelectMode) {
+      setIsMultiSelectMode(true)
+      setSelectedChores(new Set([choreId]))
+      return
+    }
+    toggleChoreSelection(choreId)
+  }
+
   const selectAllVisibleChores = () => {
     if (finalChores.length > 0) {
       setSelectedChores(new Set(finalChores.map(c => c.id)))
@@ -1051,6 +1061,7 @@ const ArchivedTasks = () => {
               isMultiSelectMode={isMultiSelectMode}
               selectedChores={selectedChores}
               toggleChoreSelection={toggleChoreSelection}
+              onLongPressChore={enterMultiSelectWithChore}
             />
           </List>
         </Box>
