+++ src/views/Chores/ChoreListView.jsx
@@ -18,8 +18,59 @@ import {
 } from '@mui/icons-material'
 import { Box, Typography } from '@mui/joy'
 import { useNavigate } from 'react-router-dom'
+import { useLongPress } from '../../hooks/useLongPress'
 import ChoreCard from './ChoreCard'
 import CompactChoreCard from './CompactChoreCard'
+
+/**
+ * One swipeable row. Owns the press-and-hold gesture (multi-select), which
+ * can't live in the render loop because it needs a hook.
+ */
+const ChoreSwipeableItem = ({
+  trailingActions,
+  onClick,
+  onLongPress,
+  longPressEnabled,
+  children,
+  // SwipeableList clones its children to inject list-level config
+  // (listType, fullSwipe, thresholds…), so it has to be passed through.
+  ...listProps
+}) => {
+  const { handlers: longPressHandlers, cancel: cancelLongPress } = useLongPress(
+    onLongPress,
+    { enabled: longPressEnabled },
+  )
+
+  // The swipe list owns the gesture the moment it recognizes a drag — a hold
+  // that turned into a swipe must not also open multi-select.
+  const handleSwipeStart = () => {
+    cancelLongPress()
+  }
+
+  return (
+    <SwipeableListItem
+      {...listProps}
+      trailingActions={trailingActions}
+      onClick={onClick}
+      onSwipeStart={handleSwipeStart}
+      onSwipeProgress={cancelLongPress}
+    >
+      <Box
+        {...longPressHandlers}
+        sx={{
+          width: '100%',
+          // Keep a long press from selecting the task text / popping the
+          // native callout on mobile
+          userSelect: 'none',
+          WebkitUserSelect: 'none',
+          WebkitTouchCallout: 'none',
+        }}
+      >
+        {children}
+      </Box>
+    </SwipeableListItem>
+  )
+}
 
 const ChoreListView = ({
   chores,
@@ -34,6 +85,7 @@ const ChoreListView = ({
   userProfile,
   isOfficialInstance,
   toggleMultiSelectMode,
+  onLongPressChore,
   showActions = true,
 }) => {
   const navigate = useNavigate()
@@ -248,7 +300,7 @@ const ChoreListView = ({
     return (
       <SwipeableList type={ListType.IOS} fullSwipe={false}>
         {chores.map(chore => (
-          <SwipeableListItem
+          <ChoreSwipeableItem
             key={chore.id}
             trailingActions={getTrailingActions(chore)}
             onClick={() => {
@@ -258,9 +310,11 @@ const ChoreListView = ({
                 navigate(`/chores/${chore.id}`)
               }
             }}
+            longPressEnabled={Boolean(onLongPressChore)}
+            onLongPress={() => onLongPressChore?.(chore.id)}
           >
             {renderChoreCard(chore)}
-          </SwipeableListItem>
+          </ChoreSwipeableItem>
         ))}
       </SwipeableList>
     )
